requirements.txt
httpx==0.27.*
tenacity==8.3.*
beautifulsoup4==4.12.*
lxml==5.*
feedparser==6.0.*
python-dateutil==2.9.*
Jinja2==3.1.*
rapidfuzz==3.9.*
pyYAML==6.0.*
